pyproject.toml
[project]
name = "s-trans"
version = "0.1.0"
description = "Document translator web app — preserves styling and supports RTL"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.5",
  "python-multipart>=0.0.12",
  "litellm>=1.50",
  "tenacity>=9.0",
  "python-docx>=1.1",
  "python-pptx>=1.0",
  "openpyxl>=3.1",
  "pymupdf>=1.24",
  "pypdf>=5.0",
  "docxcompose>=1.4",
  "fast-langdetect>=0.2",
  "arabic-reshaper>=3.0",
  "python-bidi>=0.6",
  "lxml>=5.3",
  "tiktoken>=0.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
